Fixed checkstyles, PMD and tests
