+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/artefactsummary/CopDailyCauseList.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.pip.publication.services.service.helpers.DataManipulation;
-import uk.gov.hmcts.reform.pip.publication.services.service.helpers.Helpers;
+import uk.gov.hmcts.reform.pip.publication.services.service.helpers.GeneralHelper;
 
 @Service
 public class CopDailyCauseList {
@@ -32,33 +32,28 @@ public class CopDailyCauseList {
      * @return String containing the summary for the list.
      */
     private String processCopDailyCauseList(JsonNode node) {
-        StringBuilder output = new StringBuilder();
+        StringBuilder output = new StringBuilder(100);
         node.get("courtLists").forEach(courtList -> {
             courtList.get("courtHouse").get("courtRoom").forEach(courtRoom -> {
                 courtRoom.get("session").forEach(session -> {
                     session.get("sittings").forEach(sitting -> {
                         sitting.get("hearing").forEach(hearing -> {
                             hearing.get("case").forEach(hearingCase -> {
-                                output.append('\n').append('\n')
-                                    .append("Name of Party(ies) - ")
-                                    .append(Helpers.findAndReturnNodeText(hearingCase, "caseSuppressionName"))
-                                    .append('\n')
-                                    .append("Case ID - ")
-                                    .append(Helpers.findAndReturnNodeText(hearingCase, "caseNumber"))
-                                    .append('\n').append('\n')
-                                    .append("Hearing Type - ")
-                                    .append(Helpers.findAndReturnNodeText(hearing, "hearingType"))
-                                    .append('\n')
-                                    .append("Location - ")
-                                    .append(Helpers.findAndReturnNodeText(sitting, "caseHearingChannel"))
-                                    .append('\n')
-                                    .append("Duration - ")
-                                    .append(Helpers.findAndReturnNodeText(sitting, "formattedDuration"))
-                                    .append(" ")
-                                    .append(Helpers.findAndReturnNodeText(hearingCase, "caseIndicator"))
-                                    .append('\n')
-                                    .append("Before Hon - ")
-                                    .append(Helpers.findAndReturnNodeText(session, "formattedSessionJoh"));
+                                output
+                                    .append("\n\nName of Party(ies) - ")
+                                    .append(GeneralHelper.findAndReturnNodeText(hearingCase, "caseSuppressionName"))
+                                    .append("\nCase ID - ")
+                                    .append(GeneralHelper.findAndReturnNodeText(hearingCase, "caseNumber"))
+                                    .append("\nHearing Type - ")
+                                    .append(GeneralHelper.findAndReturnNodeText(hearing, "hearingType"))
+                                    .append("\nLocation - ")
+                                    .append(GeneralHelper.findAndReturnNodeText(sitting, "caseHearingChannel"))
+                                    .append("\nDuration - ")
+                                    .append(GeneralHelper.findAndReturnNodeText(sitting, "formattedDuration"))
+                                    .append(' ')
+                                    .append(GeneralHelper.findAndReturnNodeText(hearingCase, "caseIndicator"))
+                                    .append("\nBefore Hon - ")
+                                    .append(GeneralHelper.findAndReturnNodeText(session, "formattedSessionJoh"));
                             });
                         });
                     });
