+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/helpers/DateHelper.java
@@ -1,22 +1,16 @@
 package uk.gov.hmcts.reform.pip.publication.services.service.helpers;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import org.apache.commons.lang3.StringUtils;
-
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 
-/**
- * Class for static utility methods assisting with json->html->pdf issues.
- */
-public final class Helpers {
+public final class DateHelper {
 
     private static final int ONE = 1;
 
-    private Helpers() {
+    private DateHelper() {
         throw new UnsupportedOperationException();
     }
 
@@ -54,21 +48,6 @@ public final class Helpers {
         return unZonedDateTime.atZone(zone);
     }
 
-    public static String stringDelimiter(String text, String delimiter) {
-        return text.isEmpty() ? "" : delimiter;
-    }
-
-    public static String findAndReturnNodeText(JsonNode node, String nodeName) {
-        if (node.has(nodeName)) {
-            return node.get(nodeName).asText();
-        }
-        return "";
-    }
-
-    public static String trimAnyCharacterFromStringEnd(String text) {
-        return StringUtils.isBlank(text) ? "" : text.trim().replaceAll(",$", "");
-    }
-
     public static int convertTimeToMinutes(ZonedDateTime startDateTime,
                                            ZonedDateTime endDateTime) {
         int diffHours = endDateTime.getHour() - startDateTime.getHour();
@@ -102,17 +81,6 @@ public final class Helpers {
         Instant unZonedDateTime = Instant.parse(timestamp);
         ZoneId zone = ZoneId.of("Europe/London");
         return unZonedDateTime.atZone(zone);
-    }
-
-    public static void loopAndFormatString(JsonNode nodes, String nodeName,
-                                           StringBuilder builder, String delimiter) {
-        nodes.get(nodeName).forEach(node -> {
-            if (!node.asText().isEmpty()) {
-                builder
-                    .append(node.asText())
-                    .append(delimiter);
-            }
-        });
     }
 
 }
