+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/service/artefactsummary/CopDailyCauseListTest.java
@@ -16,7 +16,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @ActiveProfiles("test")
-public class CopDailyCauseListTest {
+class CopDailyCauseListTest {
 
     @Autowired
     CopDailyCauseList copDailyCauseList;
