+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/service/pdf/converters/CopCauseListConverterTest.java
@@ -22,7 +22,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
 @ActiveProfiles("test")
-public class CopCauseListConverterTest {
+class CopCauseListConverterTest {
 
     @Autowired
     CopDailyCauseListConverter copDailyCauseListConverter;
